back/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
cryptography==43.0.3
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-classy-tags==4.1.0
django-cors-headers==4.6.0
django-environ==0.12.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-nyt==1.4.2
django-sekizai==4.1.0
djangorestframework==3.15.2
exceptiongroup==1.3.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
joblib==1.5.1
Markdown==3.6
numpy==2.0.2
oauthlib==3.2.2
openai==1.82.0
pillow==11.2.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
pymdown-extensions==10.5
python-dotenv==1.1.0
python-slugify==8.0.4
python3-openid==3.2.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.6.1
scipy==1.13.1
sniffio==1.3.1
sorl-thumbnail==12.11.0
soupsieve==2.7
sqlparse==0.5.1
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
wiki==0.12.0
wikipedia==1.4.0
Wikipedia-API==0.8.1
zipp==3.21.0
